refactor(mcp): split server.go into thematic tool files
Move tool registrations from server.go (~1170 lines) into smaller
files grouped by purpose. No behavior change — pure code move.

server.go now keeps only NewServer + serverInstructions (94 lines).
New files: tools_query.go, tools_edit.go, tools_interface.go,
tools_insert_call.go, tools_plan.go, tools_codegen.go, tools_patch.go.

registerOtherTools is split into registerExecutePlanTool (plan)
and registerCodegenTools (implement, mock, test, tag,
extract_interface).
