+++ internal/surgeon/inbound/mcp/server.go
@@ -1,11 +1,6 @@
 package mcp
 
 import (
-	"context"
-	"fmt"
-	"strings"
-
-	"github.com/JLugagne/go-surgeon/internal/surgeon/domain"
 	"github.com/JLugagne/go-surgeon/internal/surgeon/domain/service"
 	"github.com/modelcontextprotocol/go-sdk/mcp"
 )
@@ -85,7 +80,8 @@ func NewServer(commands service.SurgeonCommands, queries service.SurgeonQueries)
 	registerActionTools(s, commands)
 	registerInterfaceTools(s, commands)
 	registerInsertCallTool(s, commands)
-	registerOtherTools(s, commands)
+	registerExecutePlanTool(s, commands)
+	registerCodegenTools(s, commands)
 	registerPatchTools(s, commands)
 	registerReferencesTools(s, queries)
 	registerBatchQueryTool(s, queries)
@@ -95,1077 +91,4 @@ func NewServer(commands service.SurgeonCommands, queries service.SurgeonQueries)
 	s.AddReceivingMiddleware(schemaHintMiddleware())
 
 	return s
-}
-
-// --- Query tools ---
-
-type graphInput struct {
-	Dir         string   `json:"dir,omitempty" jsonschema:"directory to walk, defaults to current directory"`
-	Symbols     bool     `json:"symbols,omitempty" jsonschema:"include exported symbols per file"`
-	Summary     bool     `json:"summary,omitempty" jsonschema:"append package doc comment summary"`
-	Deps        bool     `json:"deps,omitempty" jsonschema:"show internal package import dependencies"`
-	Recursive   bool     `json:"recursive,omitempty" jsonschema:"walk sub-packages when symbols is set"`
-	Tests       bool     `json:"tests,omitempty" jsonschema:"include test files"`
-	Depth       int      `json:"depth,omitempty" jsonschema:"limit directory recursion depth, 0 means unlimited"`
-	Focus       string   `json:"focus,omitempty" jsonschema:"package path for full detail, others show path only"`
-	Exclude     []string `json:"exclude,omitempty" jsonschema:"glob patterns for directories to skip"`
-	TokenBudget int      `json:"token_budget,omitempty" jsonschema:"approximate max tokens in output, 0 means unlimited"`
-	Module      string   `json:"module,omitempty" jsonschema:"import path of a dependency to explore instead of the current project, e.g. 'github.com/spf13/cobra'; dir and focus are relative to the module root when set"`
-}
-
-type buildCheckInput struct {
-	Dir            string `json:"dir,omitempty" jsonschema:"relative directory or package pattern to build, defaults to './...'. Absolute paths and '..' traversal are rejected."`
-	Tests          bool   `json:"tests,omitempty" jsonschema:"compile test files too (uses 'go test -count=0 -run ^$' instead of 'go build')"`
-	TimeoutSeconds int    `json:"timeout_seconds,omitempty" jsonschema:"timeout in seconds, default 60, max 600"`
-	AffectedBy     string `json:"affected_by,omitempty" jsonschema:"path to a .go file — narrow the build to the package that owns this file plus every package in the module that (transitively) imports it. Mutually exclusive with dir. Great after editing one file in a large monorepo — skips rebuilding unrelated packages."`
-}
-
-type symbolInput struct {
-	Query       string `json:"query,omitempty" jsonschema:"exact symbol name: Name, Receiver.Method, or pkg.Name. Mutually exclusive with pattern."`
-	Pattern     string `json:"pattern,omitempty" jsonschema:"regex to match against declaration names (funcs, methods, types); returns the list of matches. Mutually exclusive with query."`
-	Body        bool   `json:"body,omitempty" jsonschema:"show the full function or struct body with line numbers"`
-	Tests       bool   `json:"tests,omitempty" jsonschema:"include test files in the search"`
-	Dir         string `json:"dir,omitempty" jsonschema:"directory to search in, defaults to current directory"`
-	Module      string `json:"module,omitempty" jsonschema:"import path of a dependency to search in instead of the current project, e.g. 'github.com/spf13/cobra'"`
-	Context     string `json:"context,omitempty" jsonschema:"optional 'file' to additionally return an outline of sibling declarations in the same file (signatures only, with line ranges) — set this when you might edit nearby code and want to see the file's shape in one call"`
-	TokenBudget int    `json:"token_budget,omitempty" jsonschema:"approximate max tokens in output (pattern mode), 0 means unlimited"`
-	Outline     bool   `json:"outline,omitempty" jsonschema:"pattern mode only: include first-sentence doc summary per match (middle ground between signature-only and body=true)"`
-	File        string `json:"file,omitempty" jsonschema:"absolute or relative path to the Go file; required when at_line is set"`
-	AtLine      int    `json:"at_line,omitempty" jsonschema:"resolve the outermost declaration that spans this 1-based line number in file; mutually exclusive with query and pattern"`
-	MaxResults  int    `json:"max_results,omitempty" jsonschema:"pattern mode only: cap the number of results returned; 0 = unlimited (default). Use to avoid large payloads with broad patterns."`
-}
-
-func registerQueryTools(s *mcp.Server, queries service.SurgeonQueries) {
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "overview",
-		Description: "List packages and symbols across the project. START HERE on an unfamiliar codebase. Also use when entering a new package: focus=pkg/path symbols=true shows every type/func/interface in one call. Explore dependencies with module='github.com/org/repo'. token_budget caps output on large projects.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in graphInput) (*mcp.CallToolResult, any, error) {
-		dir := in.Dir
-		if dir == "" {
-			dir = "."
-		}
-
-		opts := domain.GraphOptions{
-			Dir:         dir,
-			Symbols:     in.Symbols,
-			Summary:     in.Summary,
-			Deps:        in.Deps,
-			Recursive:   in.Recursive,
-			Tests:       in.Tests,
-			Depth:       in.Depth,
-			Focus:       in.Focus,
-			Exclude:     in.Exclude,
-			TokenBudget: in.TokenBudget,
-			Module:      in.Module,
-		}
-
-		if opts.Focus != "" {
-			opts.Symbols = true
-			opts.Summary = true
-			opts.Recursive = true
-		}
-
-		packages, err := queries.Graph(ctx, opts)
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("failed to build graph: %v", err), err), nil, nil
-		}
-
-		text := formatGraph(packages, opts)
-		return textResult(text), nil, nil
-	})
-
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "symbol",
-		Description: "Read one declaration (exact query='Name'/'Receiver.Method'/'pkg.Name') or list matches (pattern=regex). To resolve a file:line diagnostic directly, set file+at_line — returns the outermost named declaration that spans that line, no name lookup needed. body=true shows the implementation plus the file's package line and import block. In pattern mode, outline=true returns signature + first-sentence doc summary per match (middle ground between signature-only and body=true). When exploring an unfamiliar file, use context=file to also get an outline of every sibling declaration — saves 5+ follow-up calls. Works on dependencies via module='github.com/org/repo'. query, pattern, and at_line are mutually exclusive.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in symbolInput) (*mcp.CallToolResult, any, error) {
-		dir := in.Dir
-		if dir == "" {
-			dir = "."
-		}
-
-		atLineSet := in.AtLine > 0
-		if atLineSet && (in.Query != "" || in.Pattern != "") {
-			return errorResult("at_line is mutually exclusive with query and pattern"), nil, nil
-		}
-		if in.Query != "" && in.Pattern != "" {
-			return errorResult("query and pattern are mutually exclusive — set one"), nil, nil
-		}
-		if !atLineSet && in.Query == "" && in.Pattern == "" {
-			return errorResult("one of query, pattern, or at_line (with file) is required"), nil, nil
-		}
-		if atLineSet && in.File == "" {
-			return errorResult("file is required when at_line is set"), nil, nil
-		}
-
-		if atLineSet {
-			results, err := queries.FindSymbols(ctx, domain.SymbolQuery{File: in.File, AtLine: in.AtLine, Context: in.Context, Tests: in.Tests}, dir)
-			if err != nil {
-				return errorResultWithCode(err.Error(), err), nil, nil
-			}
-			if len(results) == 0 {
-				hint := fmt.Sprintf("Line %d is not a named declaration (it may be a package clause, import block, blank line, or comment). Try a line inside a func/type/const/var body, or use query='SymbolName' instead.", in.AtLine)
-				return textResult(fmt.Sprintf("No declaration found at %s:%d.\n%s", in.File, in.AtLine, hint)), nil, nil
-			}
-			text := formatSymbolResults(results, in.Body, fmt.Sprintf("%s:%d", in.File, in.AtLine))
-			return textResult(text), nil, nil
-		}
-
-		if in.Pattern != "" {
-			results, err := queries.FindSymbols(ctx, domain.SymbolQuery{Pattern: in.Pattern, Tests: in.Tests, Module: in.Module, Context: in.Context, MaxResults: in.MaxResults}, dir)
-			if err != nil {
-				return errorResultWithCode(err.Error(), err), nil, nil
-			}
-			if len(results) == 0 {
-				return textResult(fmt.Sprintf("No declarations match pattern %q.", in.Pattern)), nil, nil
-			}
-			text := formatPatternResults(results, in.Body, in.Pattern, in.TokenBudget, in.Outline)
-			return textResult(text), nil, nil
-		}
-
-		results := findSymbols(ctx, queries, in.Query, in.Tests, dir, in.Module, in.Context)
-		if len(results) == 0 {
-			return textResult(fmt.Sprintf("No matches found for '%s'.\nHint: run 'overview' with symbols=true and dir set to list available symbols.", in.Query)), nil, nil
-		}
-
-		text := formatSymbolResults(results, in.Body, in.Query)
-		return textResult(text), nil, nil
-	})
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "build_check",
-		Description: "Run `go build` against a package/directory and return structured compile diagnostics. Use this AFTER editing Go code to verify the package still compiles — diagnostics carry file:line:column + message, deduplicated per file. Set tests=true to also compile _test.go files. dir defaults to './...' (the whole module); pass a relative path like 'internal/foo' or 'internal/foo/...' to scope the check. timeout_seconds defaults to 60 (max 600). `go vet` is out of scope — this tool only reports errors the compiler itself sees.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in buildCheckInput) (*mcp.CallToolResult, any, error) {
-		result, err := queries.BuildCheck(ctx, domain.BuildCheckRequest{
-			Dir:            in.Dir,
-			Tests:          in.Tests,
-			TimeoutSeconds: in.TimeoutSeconds,
-			AffectedBy:     in.AffectedBy,
-		})
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("build_check failed: %v", err), err), nil, nil
-		}
-
-		text := formatBuildCheckResult(result)
-		res := textResult(text)
-		res.StructuredContent = buildCheckOutput{
-			Success:     result.Success,
-			Diagnostics: convertBuildDiagnostics(result.Diagnostics),
-			RawOutput:   result.RawOutput,
-			ExitCode:    result.ExitCode,
-			DurationMs:  result.DurationMs,
-			TimedOut:    result.TimedOut,
-			Truncated:   result.Truncated,
-		}
-		return res, nil, nil
-	})
-
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "test_run",
-		Description: "Run `go test` scoped to a package/directory and return a compact pass/fail report with per-test timing and failure file:line references. Use after editing Go code to verify behavior in-loop. dir defaults to ./..., timeout defaults to 120s (max 600). Pass affected_by=path/to/file.go to run only the owning package plus its reverse-dependency closure (mutually exclusive with dir and symbols). Pass symbols=['pkg.MyFunc'] to auto-resolve the owning package and build a -run filter matching ^TestMyFunc — ideal when you only want tests related to specific functions you just edited; mutually exclusive with dir and affected_by. On success the verbatim raw_output stream is omitted from the structured payload (it bloats responses without adding signal); set include_raw_output=true to force it on a green run. verbosity controls payload size on large suites: 'summary' returns only success/summary/failed tests (~1k chars regardless of suite size — ideal for the 25k token tool-result budget); 'full' keeps everything; default auto picks 'summary' once the suite has more than 50 tests.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in testRunInput) (*mcp.CallToolResult, any, error) {
-		result, err := queries.TestRun(ctx, domain.TestRunRequest{
-			Dir:            in.Dir,
-			Run:            in.Run,
-			Count:          in.Count,
-			Race:           in.Race,
-			Tags:           in.Tags,
-			TimeoutSeconds: in.TimeoutSeconds,
-			AffectedBy:     in.AffectedBy,
-			Symbols:        in.Symbols,
-		})
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("ERROR (test_run): %v", err), err), nil, nil
-		}
-		if result.Success && !in.IncludeRawOutput {
-			result.RawOutput = ""
-		}
-		structured := applyTestRunVerbosity(result, in.Verbosity, in.IncludeRawOutput)
-		res := textResult(formatTestRunResult(result))
-		res.StructuredContent = structured
-		return res, nil, nil
-	})
-}
-
-// --- Action tools (create / update / delete) ---
-
-type createInput struct {
-	Object     string `json:"object" jsonschema:"what to create: file, func, or struct"`
-	File       string `json:"file" jsonschema:"target file path"`
-	Content    string `json:"content" jsonschema:"raw Go source code, no package declaration or imports"`
-	Identifier string `json:"identifier,omitempty" jsonschema:"ignored — identifier is inferred from content; accepted to avoid validation errors"`
-	WithTest   bool   `json:"with_test,omitempty" jsonschema:"generate a test skeleton alongside the function (only applies when object=func)"`
-	Preview    bool   `json:"preview,omitempty" jsonschema:"if true, return diff without writing the file"`
-}
-
-type updateInput struct {
-	Object     string `json:"object" jsonschema:"what to update: file, func, or struct"`
-	File       string `json:"file" jsonschema:"target file path"`
-	Identifier string `json:"identifier,omitempty" jsonschema:"AST identifier, e.g. FuncName or Receiver.Method, required for func and struct"`
-	Content    string `json:"content" jsonschema:"raw Go source code, no package declaration or imports"`
-	Doc        string `json:"doc,omitempty" jsonschema:"set or replace the doc comment (raw text without // prefix)"`
-	StripDoc   bool   `json:"strip_doc,omitempty" jsonschema:"remove the existing doc comment"`
-	WithTest   bool   `json:"with_test,omitempty" jsonschema:"generate a test skeleton alongside the function (only applies when object=func)"`
-	Preview    bool   `json:"preview,omitempty" jsonschema:"if true, return diff without writing the file"`
-}
-
-type deleteInput struct {
-	Object     string `json:"object" jsonschema:"what to delete: func or struct"`
-	File       string `json:"file" jsonschema:"target file path"`
-	Identifier string `json:"identifier,omitempty" jsonschema:"AST identifier (e.g. FuncName or Receiver.Method); not required when object=file"`
-	Preview    bool   `json:"preview,omitempty" jsonschema:"if true, return diff without writing the file"`
-}
-
-var createObjectMap = map[string]domain.ActionType{
-	"file":   domain.ActionTypeCreateFile,
-	"func":   domain.ActionTypeAddFunc,
-	"struct": domain.ActionTypeAddStruct,
-}
-
-var updateObjectMap = map[string]domain.ActionType{
-	"file":   domain.ActionTypeReplaceFile,
-	"func":   domain.ActionTypeUpdateFunc,
-	"struct": domain.ActionTypeUpdateStruct,
-}
-
-var deleteObjectMap = map[string]domain.ActionType{
-	"func":   domain.ActionTypeDeleteFunc,
-	"struct": domain.ActionTypeDeleteStruct,
-	"file":   domain.ActionTypeDeleteFile,
-}
-
-func registerActionTools(s *mcp.Server, commands service.SurgeonCommands) {
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "create",
-		Description: "Add a new file, function, or struct. object='file' creates a new file; 'func'/'struct' append to an existing file; 'auto' infers from content (func ... → func, type ... struct → struct, else file). When adding several related items, use execute_plan instead.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in createInput) (*mcp.CallToolResult, any, error) {
-		if err := validateGoFile(in.File); err != nil {
-			return err, nil, nil
-		}
-		var actionType domain.ActionType
-		if in.Object == "auto" {
-			trimmed := strings.TrimSpace(in.Content)
-			if strings.HasPrefix(trimmed, "func ") {
-				actionType = domain.ActionTypeAddFunc
-			} else if strings.Contains(trimmed, "type ") && strings.Contains(trimmed, "struct {") {
-				actionType = domain.ActionTypeAddStruct
-			} else {
-				actionType = domain.ActionTypeCreateFile
-			}
-		} else {
-			var ok bool
-			actionType, ok = createObjectMap[in.Object]
-			if !ok {
-				return errorResult(fmt.Sprintf("invalid object %q: must be file, func, struct, or auto", in.Object)), nil, nil
-			}
-		}
-
-		result, err := commands.ExecutePlan(ctx, domain.Plan{
-			Preview: in.Preview,
-			Actions: []domain.Action{{
-				Action:   actionType,
-				FilePath: in.File,
-				Content:  in.Content,
-				WithTest: in.WithTest,
-			}},
-		})
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("ERROR (create %s): %v", in.Object, err), err), nil, nil
-		}
-
-		var sb strings.Builder
-		for _, w := range result.Warnings {
-			fmt.Fprintf(&sb, "WARNING: %s\n", w)
-		}
-		verb := "SUCCESS"
-		if result.Preview {
-			verb = "PREVIEW"
-		}
-		fmt.Fprintf(&sb, "%s (create %s): %d files modified", verb, in.Object, result.FilesModified)
-		if result.Diff != "" {
-			sb.WriteString("\n\n")
-			sb.WriteString(result.Diff)
-		}
-		res := textResult(sb.String())
-		res.StructuredContent = editOutput{
-			FilesModified: result.Files,
-			Symbols:       []symbolEdit{{Action: string(actionType), File: in.File}},
-			Warnings:      result.Warnings,
-			AddedImports:  result.AddedImports,
-			Preview:       result.Preview,
-			Diff:          result.Diff,
-		}
-		return res, nil, nil
-	})
-
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "update",
-		Description: "Replace a whole function, method, struct, or file. For small changes inside a function body, prefer patch_function. content is the complete new declaration (signature + body). Doc comments are kept unless you set doc or strip_doc=true. preview=true returns a unified diff without writing. object='auto' infers from content (func ... → func, type ... struct → struct, else file).",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in updateInput) (*mcp.CallToolResult, any, error) {
-		if err := validateGoFile(in.File); err != nil {
-			return err, nil, nil
-		}
-		var actionType domain.ActionType
-		if in.Object == "auto" {
-			trimmed := strings.TrimSpace(in.Content)
-			if strings.HasPrefix(trimmed, "func ") {
-				actionType = domain.ActionTypeUpdateFunc
-			} else if strings.Contains(trimmed, "type ") && strings.Contains(trimmed, "struct {") {
-				actionType = domain.ActionTypeUpdateStruct
-			} else {
-				actionType = domain.ActionTypeReplaceFile
-			}
-		} else {
-			var ok bool
-			actionType, ok = updateObjectMap[in.Object]
-			if !ok {
-				return errorResult(fmt.Sprintf("invalid object %q: must be file, func, struct, or auto", in.Object)), nil, nil
-			}
-		}
-
-		result, err := commands.ExecutePlan(ctx, domain.Plan{
-			Preview: in.Preview,
-			Actions: []domain.Action{{
-				Action:     actionType,
-				FilePath:   in.File,
-				Identifier: in.Identifier,
-				Content:    in.Content,
-				Doc:        in.Doc,
-				StripDoc:   in.StripDoc,
-				WithTest:   in.WithTest,
-			}},
-		})
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("ERROR (update %s): %v", in.Object, err), err), nil, nil
-		}
-
-		var sb strings.Builder
-		for _, w := range result.Warnings {
-			fmt.Fprintf(&sb, "WARNING: %s\n", w)
-		}
-		verb := "SUCCESS"
-		if result.Preview {
-			verb = "PREVIEW"
-		}
-		fmt.Fprintf(&sb, "%s (update %s): %d files modified", verb, in.Object, result.FilesModified)
-		if result.Diff != "" {
-			sb.WriteString("\n\n")
-			sb.WriteString(result.Diff)
-		}
-		res := textResult(sb.String())
-		res.StructuredContent = editOutput{
-			FilesModified: result.Files,
-			Symbols:       []symbolEdit{{Action: string(actionType), Identifier: in.Identifier, File: in.File}},
-			Warnings:      result.Warnings,
-			AddedImports:  result.AddedImports,
-			Preview:       result.Preview,
-			Diff:          result.Diff,
-		}
-		return res, nil, nil
-	})
-
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "delete",
-		Description: "Remove a function, method, struct, or file. object='file' deletes the file from disk (no identifier needed); object='struct' also removes every method on the struct across the package. For interfaces, use delete_interface instead (handles mock cleanup). preview=true returns a unified diff without writing.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in deleteInput) (*mcp.CallToolResult, any, error) {
-		if err := validateGoFile(in.File); err != nil {
-			return err, nil, nil
-		}
-		actionType, ok := deleteObjectMap[in.Object]
-		if !ok {
-			return errorResult(fmt.Sprintf("invalid object %q: must be file, func, or struct", in.Object)), nil, nil
-		}
-
-		result, err := commands.ExecutePlan(ctx, domain.Plan{
-			Preview: in.Preview,
-			Actions: []domain.Action{{
-				Action:     actionType,
-				FilePath:   in.File,
-				Identifier: in.Identifier,
-			}},
-		})
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("ERROR (delete %s): %v", in.Object, err), err), nil, nil
-		}
-
-		var sb strings.Builder
-		for _, w := range result.Warnings {
-			fmt.Fprintf(&sb, "WARNING: %s\n", w)
-		}
-		verb := "SUCCESS"
-		if result.Preview {
-			verb = "PREVIEW"
-		}
-		fmt.Fprintf(&sb, "%s (delete %s): %d files modified", verb, in.Object, result.FilesModified)
-		if result.Diff != "" {
-			sb.WriteString("\n\n")
-			sb.WriteString(result.Diff)
-		}
-		res := textResult(sb.String())
-		res.StructuredContent = editOutput{
-			FilesModified: result.Files,
-			Symbols:       []symbolEdit{{Action: string(actionType), Identifier: in.Identifier, File: in.File}},
-			Warnings:      result.Warnings,
-			AddedImports:  result.AddedImports,
-			Preview:       result.Preview,
-			Diff:          result.Diff,
-		}
-		return res, nil, nil
-	})
-}
-
-// --- Interface tools ---
-
-type interfaceInput struct {
-	File       string `json:"file" jsonschema:"file containing the interface, required"`
-	Identifier string `json:"identifier,omitempty" jsonschema:"interface name, required for update and delete"`
-	Content    string `json:"content,omitempty" jsonschema:"raw Go interface source, no package declaration or imports"`
-	MockFile   string `json:"mock_file,omitempty" jsonschema:"target file for the generated mock (add/update), or file containing the mock to delete (delete)"`
-	MockName   string `json:"mock_name,omitempty" jsonschema:"name of the mock struct"`
-	Doc        string `json:"doc,omitempty" jsonschema:"set or replace the doc comment (raw text without // prefix, update only)"`
-	StripDoc   bool   `json:"strip_doc,omitempty" jsonschema:"remove the existing doc comment (update only)"`
-	DeleteMock bool   `json:"delete_mock,omitempty" jsonschema:"delete only: also remove the mock struct, its methods and its compile-time assertion from mock_file; requires mock_file and mock_name"`
-	Preview    bool   `json:"preview,omitempty" jsonschema:"if true, return diff without writing the file"`
-}
-
-func registerInterfaceTools(s *mcp.Server, commands service.SurgeonCommands) {
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "add_interface",
-		Description: "Add a new interface and its mock atomically. Set mock_file + mock_name to generate a function-field mock with a compile-time var assertion. Prefer this over create for interfaces. preview=true returns a unified diff without writing.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in interfaceInput) (*mcp.CallToolResult, any, error) {
-		if err := validateGoFile(in.File); err != nil {
-			return err, nil, nil
-		}
-		reqDomain := domain.InterfaceActionRequest{
-			FilePath: in.File, Identifier: in.Identifier, Content: in.Content,
-			MockFile: in.MockFile, MockName: in.MockName,
-		}
-		var result string
-		var addedImports []string
-		var diff string
-		var writtenFiles []string
-		var err error
-		if in.Preview {
-			diff, writtenFiles, err = runPreview(ctx, commands, func(sc service.SurgeonCommands) error {
-				var innerErr error
-				result, addedImports, innerErr = sc.AddInterface(ctx, reqDomain)
-				return innerErr
-			})
-		} else {
-			result, addedImports, err = commands.AddInterface(ctx, reqDomain)
-		}
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("ERROR (add_interface): %v", err), err), nil, nil
-		}
-		files := []string{in.File}
-		if in.MockFile != "" {
-			files = append(files, in.MockFile)
-		}
-		if in.Preview && len(writtenFiles) > 0 {
-			files = writtenFiles
-		}
-		msg := result
-		if in.Preview {
-			msg = fmt.Sprintf("PREVIEW (add_interface): %s", result)
-			if diff != "" {
-				msg += "\n\n" + diff
-			}
-		}
-		res := textResult(msg)
-		res.StructuredContent = editOutput{
-			FilesModified: files,
-			Symbols:       []symbolEdit{{Action: "add_interface", Identifier: in.Identifier, File: in.File}},
-			AddedImports:  addedImports,
-			Preview:       in.Preview,
-			Diff:          diff,
-		}
-		return res, nil, nil
-	})
-
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "update_interface",
-		Description: "Replace an interface wholesale and regenerate its mock. For single-method changes, prefer patch_interface. Pass mock_file + mock_name to keep the mock in sync. Doc comments are kept unless you set doc or strip_doc=true. preview=true returns a unified diff without writing.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in interfaceInput) (*mcp.CallToolResult, any, error) {
-		if err := validateGoFile(in.File); err != nil {
-			return err, nil, nil
-		}
-		reqDomain := domain.InterfaceActionRequest{
-			FilePath: in.File, Identifier: in.Identifier, Content: in.Content,
-			MockFile: in.MockFile, MockName: in.MockName,
-			Doc: in.Doc, StripDoc: in.StripDoc,
-		}
-		var result string
-		var addedImports []string
-		var diff string
-		var writtenFiles []string
-		var err error
-		if in.Preview {
-			diff, writtenFiles, err = runPreview(ctx, commands, func(sc service.SurgeonCommands) error {
-				var innerErr error
-				result, addedImports, innerErr = sc.UpdateInterface(ctx, reqDomain)
-				return innerErr
-			})
-		} else {
-			result, addedImports, err = commands.UpdateInterface(ctx, reqDomain)
-		}
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("ERROR (update_interface): %v", err), err), nil, nil
-		}
-		files := []string{in.File}
-		if in.MockFile != "" {
-			files = append(files, in.MockFile)
-		}
-		if in.Preview && len(writtenFiles) > 0 {
-			files = writtenFiles
-		}
-		msg := result
-		if in.Preview {
-			msg = fmt.Sprintf("PREVIEW (update_interface): %s", result)
-			if diff != "" {
-				msg += "\n\n" + diff
-			}
-		}
-		res := textResult(msg)
-		res.StructuredContent = editOutput{
-			FilesModified: files,
-			Symbols:       []symbolEdit{{Action: "update_interface", Identifier: in.Identifier, File: in.File}},
-			AddedImports:  addedImports,
-			Preview:       in.Preview,
-			Diff:          diff,
-		}
-		return res, nil, nil
-	})
-
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "delete_interface",
-		Description: "Delete an interface. Pass delete_mock=true with mock_file + mock_name to also remove the mock struct, its methods, and the var assertion. The mock file is kept even if empty. preview=true returns a unified diff without writing.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in interfaceInput) (*mcp.CallToolResult, any, error) {
-		if err := validateGoFile(in.File); err != nil {
-			return err, nil, nil
-		}
-		if in.MockFile != "" {
-			if err := validateGoFile(in.MockFile); err != nil {
-				return err, nil, nil
-			}
-		}
-		reqDomain := domain.InterfaceActionRequest{
-			FilePath:   in.File,
-			Identifier: in.Identifier,
-			MockFile:   in.MockFile,
-			MockName:   in.MockName,
-			DeleteMock: in.DeleteMock,
-		}
-		var result string
-		var addedImports []string
-		var diff string
-		var writtenFiles []string
-		var err error
-		if in.Preview {
-			diff, writtenFiles, err = runPreview(ctx, commands, func(sc service.SurgeonCommands) error {
-				var innerErr error
-				result, addedImports, innerErr = sc.DeleteInterface(ctx, reqDomain)
-				return innerErr
-			})
-		} else {
-			result, addedImports, err = commands.DeleteInterface(ctx, reqDomain)
-		}
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("ERROR (delete_interface): %v", err), err), nil, nil
-		}
-		files := []string{in.File}
-		if in.DeleteMock && in.MockFile != "" {
-			files = append(files, in.MockFile)
-		}
-		if in.Preview && len(writtenFiles) > 0 {
-			files = writtenFiles
-		}
-		msg := result
-		if in.Preview {
-			msg = fmt.Sprintf("PREVIEW (delete_interface): %s", result)
-			if diff != "" {
-				msg += "\n\n" + diff
-			}
-		}
-		res := textResult(msg)
-		res.StructuredContent = editOutput{
-			FilesModified: files,
-			Symbols:       []symbolEdit{{Action: "delete_interface", Identifier: in.Identifier, File: in.File}},
-			AddedImports:  addedImports,
-			Preview:       in.Preview,
-			Diff:          diff,
-		}
-		return res, nil, nil
-	})
-}
-
-// --- Insert-call tool ---
-
-type insertCallInput struct {
-	File     string `json:"file" jsonschema:"target Go file"`
-	Function string `json:"function" jsonschema:"function identifier: FuncName or Receiver.Method"`
-	Call     string `json:"call" jsonschema:"statement to insert, e.g. setupPayOrderRoute(mux, app)"`
-	Position string `json:"position,omitempty" jsonschema:"where to insert: before-return (default), end-of-body, or after:<marker>"`
-	Preview  bool   `json:"preview,omitempty" jsonschema:"if true, return diff without writing the file"`
-}
-
-func registerInsertCallTool(s *mcp.Server, commands service.SurgeonCommands) {
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "insert_call",
-		Description: "Insert one statement into a function body. Idempotent: skipped if already present. position: 'before-return' (default), 'end-of-body', or 'after:<marker>'.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in insertCallInput) (*mcp.CallToolResult, any, error) {
-		if err := validateGoFile(in.File); err != nil {
-			return err, nil, nil
-		}
-		pos := domain.InsertPosition(in.Position)
-		if pos == "" {
-			pos = domain.InsertBeforeReturn
-		}
-		result, err := commands.ExecutePlan(ctx, domain.Plan{
-			Preview: in.Preview,
-			Actions: []domain.Action{{
-				Action:     domain.ActionTypeInsertCall,
-				FilePath:   in.File,
-				Identifier: in.Function,
-				Content:    in.Call,
-				Position:   pos,
-			}},
-		})
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("ERROR (insert_call): %v", err), err), nil, nil
-		}
-		var sb strings.Builder
-		for _, w := range result.Warnings {
-			fmt.Fprintf(&sb, "WARNING: %s\n", w)
-		}
-		verb := "SUCCESS"
-		if result.Preview {
-			verb = "PREVIEW"
-		}
-		fmt.Fprintf(&sb, "%s (insert_call): %d files modified", verb, result.FilesModified)
-		if result.Diff != "" {
-			sb.WriteString("\n\n")
-			sb.WriteString(result.Diff)
-		}
-		res := textResult(sb.String())
-		res.StructuredContent = editOutput{
-			FilesModified: result.Files,
-			Symbols:       []symbolEdit{{Action: "insert_call", Identifier: in.Function, File: in.File}},
-			Warnings:      result.Warnings,
-			AddedImports:  result.AddedImports,
-			Preview:       result.Preview,
-			Diff:          result.Diff,
-		}
-		return res, nil, nil
-	})
-}
-
-// --- Other tools ---
-
-type executePlanInput struct {
-	Actions []executePlanActionInput `json:"actions" jsonschema:"ordered list of AST actions to execute atomically (up to 15)"`
-	Preview bool                     `json:"preview,omitempty" jsonschema:"if true, return diff without writing any files"`
-}
-
-type implementInput struct {
-	Interface string `json:"interface" jsonschema:"fully qualified interface name, e.g. io.ReadCloser or github.com/org/repo/pkg.Interface"`
-	Receiver  string `json:"receiver" jsonschema:"receiver type, e.g. *MyStruct"`
-	File      string `json:"file" jsonschema:"target file to append stubs to"`
-	Preview   bool   `json:"preview,omitempty" jsonschema:"if true, return diff without writing the file"`
-}
-
-type mockInput struct {
-	Interface string `json:"interface" jsonschema:"fully qualified interface name"`
-	MockName  string `json:"mock_name" jsonschema:"name of the mock struct, e.g. MockBookRepository"`
-	File      string `json:"file" jsonschema:"target file to write the mock to"`
-	Preview   bool   `json:"preview,omitempty" jsonschema:"if true, return diff without writing the file"`
-}
-
-type testInput struct {
-	File       string `json:"file" jsonschema:"target Go file containing the function"`
-	Identifier string `json:"identifier" jsonschema:"function or method identifier, e.g. NewApp or Book.Validate"`
-	Preview    bool   `json:"preview,omitempty" jsonschema:"if true, return diff without writing the test file"`
-}
-
-type tagInput struct {
-	File       string `json:"file" jsonschema:"target Go file containing the struct"`
-	Identifier string `json:"identifier" jsonschema:"struct identifier"`
-	Field      string `json:"field,omitempty" jsonschema:"specific field name to update"`
-	Set        string `json:"set,omitempty" jsonschema:"exact tag string to set or append"`
-	Auto       string `json:"auto,omitempty" jsonschema:"auto-generate tags for exported fields, e.g. json or bson"`
-	Preview    bool   `json:"preview,omitempty" jsonschema:"if true, return diff without writing the file"`
-}
-
-type extractInterfaceInput struct {
-	File       string `json:"file" jsonschema:"target Go file containing the struct"`
-	Identifier string `json:"identifier" jsonschema:"struct identifier"`
-	Name       string `json:"name" jsonschema:"name of the interface to create"`
-	Out        string `json:"out,omitempty" jsonschema:"output file path for the interface"`
-	MockFile   string `json:"mock_file,omitempty" jsonschema:"generate mock file path"`
-	MockName   string `json:"mock_name,omitempty" jsonschema:"name of the mock struct"`
-	Preview    bool   `json:"preview,omitempty" jsonschema:"if true, return diff without writing any file"`
-}
-
-// --- Patch tools ---
-
-type patchOpInput struct {
-	Op         string   `json:"op" jsonschema:"operation: replace, insert_before, insert_after, delete, wrap"`
-	Match      string   `json:"match,omitempty" jsonschema:"literal text to match inside the function body (whitespace-normalized)"`
-	MatchRegex string   `json:"match_regex,omitempty" jsonschema:"regex alternative to match; mutually exclusive with match"`
-	Occurrence int      `json:"occurrence,omitempty" jsonschema:"which match to operate on (1-based, default 0 = error-if-ambiguous for function). Use -1 to apply to all matches."`
-	Replace    string   `json:"replace,omitempty" jsonschema:"replacement text (for replace op)"`
-	Code       string   `json:"code,omitempty" jsonschema:"line of code to insert (for insert_before / insert_after ops)"`
-	Wrap       string   `json:"wrap,omitempty" jsonschema:"wrap template with %s as the matched text (for wrap op)"`
-	AtLine     int      `json:"at_line,omitempty" jsonschema:"target a single line by file-absolute line number (1-based, matches symbol body=true output). Mutually exclusive with match/match_regex."`
-	FromLine   int      `json:"from_line,omitempty" jsonschema:"first line of a file-absolute line range (1-based, inclusive). Pair with to_line."`
-	ToLine     int      `json:"to_line,omitempty" jsonschema:"last line of a file-absolute line range (1-based, inclusive). Must be >= from_line."`
-	Params     []string `json:"params,omitempty" jsonschema:"for set_signature: new parameter list as an array of parameter declarations (without parens), e.g. [\"ctx context.Context\", \"x int\"]. Empty keeps the current params."`
-	Returns    string   `json:"returns,omitempty" jsonschema:"for set_signature: new result list, e.g. 'error' or '([]byte, error)'. Empty keeps the current returns."`
-}
-
-func registerPatchTools(s *mcp.Server, commands service.SurgeonCommands) {
-	registerPatchTool(s, commands)
-	registerPatchBulkTools(s, commands)
-}
-
-// --- patch_file ---
-
-type filePatchOpInput struct {
-	Match        string `json:"match,omitempty" jsonschema:"literal text; all occurrences are replaced. Mutually exclusive with match_regex."`
-	MatchRegex   string `json:"match_regex,omitempty" jsonschema:"RE2 regex alternative to match; use $1, $2, ... in replace for submatch substitution. Mutually exclusive with match."`
-	Replace      string `json:"replace" jsonschema:"replacement text; supports $1/$2/... when match_regex is used"`
-	MatchLiteral bool   `json:"match_literal,omitempty" jsonschema:"when true, treats match_regex as a literal string (applies regexp.QuoteMeta). Use when the pattern contains special characters like postgres://. Mutually exclusive with match."`
-	Occurrence   int    `json:"occurrence,omitempty" jsonschema:"0 = replace all occurrences (default); N = replace only the Nth occurrence (1-based). Useful when two similar blocks need separate patches."`
-	MatchMode    string `json:"match_mode,omitempty" jsonschema:"'exact' (default) or 'normalized' — normalized collapses whitespace runs before matching, useful for aligned struct/table code"`
-}
-
-type structPatchOpInput struct {
-	Op       string `json:"op" jsonschema:"operation: add_field, remove_field, rename_field, retype_field, set_tag, set_doc"`
-	Name     string `json:"name,omitempty" jsonschema:"target field name (most ops); embed type literal (e.g. io.Reader) for embedded fields"`
-	From     string `json:"from,omitempty" jsonschema:"current field name (rename_field only)"`
-	To       string `json:"to,omitempty" jsonschema:"new field name (rename_field only)"`
-	Type     string `json:"type,omitempty" jsonschema:"field type (add_field / retype_field)"`
-	Tag      string `json:"tag,omitempty" jsonschema:"struct tag content without backticks (e.g. json:\"email,omitempty\")"`
-	Doc      string `json:"doc,omitempty" jsonschema:"doc comment text (set_doc / add_field); empty string clears the doc"`
-	Before   string `json:"before,omitempty" jsonschema:"anchor: insert before this existing field (add_field only)"`
-	After    string `json:"after,omitempty" jsonschema:"anchor: insert after this existing field (add_field only)"`
-	Position string `json:"position,omitempty" jsonschema:"first or last (add_field only); default is last"`
-}
-
-type interfacePatchOpInput struct {
-	Op        string `json:"op" jsonschema:"operation: add_method, remove_method, rename_method, retype_method, set_doc, embed, remove_embed"`
-	Name      string `json:"name,omitempty" jsonschema:"target method name (most ops)"`
-	From      string `json:"from,omitempty" jsonschema:"current method name (rename_method only)"`
-	To        string `json:"to,omitempty" jsonschema:"new method name (rename_method only)"`
-	Signature string `json:"signature,omitempty" jsonschema:"full method signature including name, e.g. 'Close() error' or 'Read(p []byte) (int, error)'"`
-	Type      string `json:"type,omitempty" jsonschema:"embedded interface type, e.g. 'io.Closer' (embed / remove_embed)"`
-	Doc       string `json:"doc,omitempty" jsonschema:"doc comment text (set_doc / add_method)"`
-	Before    string `json:"before,omitempty" jsonschema:"anchor: insert before this existing member (add_method / embed)"`
-	After     string `json:"after,omitempty" jsonschema:"anchor: insert after this existing member (add_method / embed)"`
-	Position  string `json:"position,omitempty" jsonschema:"first or last (add_method / embed); default is last"`
-}
-
-func registerOtherTools(s *mcp.Server, commands service.SurgeonCommands) {
-	mcp.AddTool(s, &mcp.Tool{
-		Name: "execute_plan",
-		Description: "Apply up to 15 related AST edits atomically — any failure rolls everything back. " +
-			"Actions: create_file, replace_file, add_func, update_func, delete_func, add_struct, update_struct, delete_struct, add_interface, update_interface, delete_interface, insert_call, patch_function, patch_struct, patch_interface, patch_file, patch_decl. " +
-			"In-place patch actions (patch_function, patch_struct, patch_interface, patch_file, patch_decl) carry their operations in dedicated fields: patch_function_ops, patch_struct_ops, patch_interface_ops, patch_file_ops, patch_decl_ops.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in executePlanInput) (*mcp.CallToolResult, any, error) {
-		actions := make([]domain.Action, len(in.Actions))
-		for i, a := range in.Actions {
-			actions[i] = domain.Action{
-				Action:            domain.ActionType(a.Action),
-				FilePath:          a.File,
-				PackagePath:       a.Package,
-				Identifier:        a.Identifier,
-				Content:           a.Content,
-				MockFile:          a.MockFile,
-				MockName:          a.MockName,
-				Doc:               a.Doc,
-				StripDoc:          a.StripDoc,
-				Position:          domain.InsertPosition(a.Position),
-				WithTest:          a.WithTest,
-				PatchFunctionOps:  toFunctionPatches(a.PatchFunctionOps),
-				PatchStructOps:    toStructPatches(a.PatchStructOps),
-				PatchInterfaceOps: toInterfacePatches(a.PatchInterfaceOps),
-				PatchFileOps:      toFilePatches(a.PatchFileOps),
-				PatchDeclOps:      toFunctionPatches(a.PatchDeclOps),
-			}
-			if actions[i].FilePath != "" {
-				if err := validateGoFile(actions[i].FilePath); err != nil {
-					return err, nil, nil
-				}
-			}
-		}
-		plan := domain.Plan{Preview: in.Preview, Actions: actions}
-
-		result, err := commands.ExecutePlan(ctx, plan)
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("plan execution failed: %v", err), err), nil, nil
-		}
-
-		var sb strings.Builder
-		for _, w := range result.Warnings {
-			fmt.Fprintf(&sb, "WARNING: %s\n", w)
-		}
-		verb := "SUCCESS"
-		if result.Preview {
-			verb = "PREVIEW"
-		}
-		fmt.Fprintf(&sb, "%s: %d files modified", verb, result.FilesModified)
-		if result.Diff != "" {
-			sb.WriteString("\n\n")
-			sb.WriteString(result.Diff)
-		}
-		symbols := make([]symbolEdit, len(in.Actions))
-		for i, a := range in.Actions {
-			symbols[i] = symbolEdit{Action: a.Action, Identifier: a.Identifier, File: a.File}
-		}
-		res := textResult(sb.String())
-		res.StructuredContent = editOutput{
-			FilesModified: result.Files,
-			Symbols:       symbols,
-			Warnings:      result.Warnings,
-			AddedImports:  result.AddedImports,
-			Preview:       result.Preview,
-			Diff:          result.Diff,
-		}
-		return res, nil, nil
-	})
-
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "implement",
-		Description: "Generate method stubs for a struct to satisfy an interface. Already-implemented methods are skipped. Stubs are marked '// TODO(go-surgeon): implement'. Interface must be fully qualified (e.g. io.ReadCloser, github.com/org/repo/pkg.Interface). preview=true returns a unified diff without writing.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in implementInput) (*mcp.CallToolResult, any, error) {
-		if err := validateGoFile(in.File); err != nil {
-			return err, nil, nil
-		}
-		reqDomain := domain.ImplementRequest{
-			Interface: in.Interface,
-			Receiver:  in.Receiver,
-			FilePath:  in.File,
-		}
-		var results []domain.SymbolResult
-		var diff string
-		var err error
-		if in.Preview {
-			diff, _, err = runPreview(ctx, commands, func(sc service.SurgeonCommands) error {
-				var innerErr error
-				results, innerErr = sc.Implement(ctx, reqDomain)
-				return innerErr
-			})
-		} else {
-			results, err = commands.Implement(ctx, reqDomain)
-		}
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("failed to implement interface: %v", err), err), nil, nil
-		}
-
-		if len(results) == 0 && diff == "" {
-			res := textResult("All methods are already implemented.")
-			res.StructuredContent = implementOutput{File: in.File, Interface: in.Interface, Receiver: in.Receiver, Stubs: []string{}}
-			return res, nil, nil
-		}
-
-		var sb strings.Builder
-		verb := "Generated"
-		if in.Preview {
-			verb = "PREVIEW — would generate"
-		}
-		fmt.Fprintf(&sb, "%s %d missing methods for %s:\n\n", verb, len(results), in.Interface)
-		for _, rr := range results {
-			fmt.Fprintf(&sb, "Symbol: %s\nReceiver: %s\nFile: %s:%d-%d\nCode:\n%s\n\n",
-				rr.Name, rr.Receiver, rr.File, rr.LineStart, rr.LineEnd, rr.Code)
-		}
-		if diff != "" {
-			sb.WriteString(diff)
-		}
-		stubs := make([]string, len(results))
-		for i, r := range results {
-			stubs[i] = r.Name
-		}
-		res := textResult(sb.String())
-		res.StructuredContent = implementOutput{File: in.File, Interface: in.Interface, Receiver: in.Receiver, Stubs: stubs}
-		return res, nil, nil
-	})
-
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "mock",
-		Description: "Generate a function-field mock for an interface you don't own (stdlib, third-party). For your own interfaces, use add_interface with mock_file instead. Interface must be fully qualified. preview=true returns a unified diff without writing.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in mockInput) (*mcp.CallToolResult, any, error) {
-		if err := validateGoFile(in.File); err != nil {
-			return err, nil, nil
-		}
-		reqDomain := domain.MockRequest{
-			Interface: in.Interface,
-			Receiver:  in.MockName,
-			FilePath:  in.File,
-		}
-		var result, diff string
-		var err error
-		if in.Preview {
-			diff, _, err = runPreview(ctx, commands, func(sc service.SurgeonCommands) error {
-				var innerErr error
-				result, innerErr = sc.Mock(ctx, reqDomain)
-				return innerErr
-			})
-		} else {
-			result, err = commands.Mock(ctx, reqDomain)
-		}
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("failed to generate mock: %v", err), err), nil, nil
-		}
-		msg := result
-		if in.Preview {
-			msg = "PREVIEW (mock): " + result
-			if diff != "" {
-				msg += "\n\n" + diff
-			}
-		}
-		res := textResult(msg)
-		res.StructuredContent = mockOutput{File: in.File, Interface: in.Interface, MockName: in.MockName}
-		return res, nil, nil
-	})
-
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "test",
-		Description: "Scaffold a table-driven _test.go for a function or method. Handles boilerplate (t.Run loop, tt struct, receiver setup). identifier: 'FuncName' or 'Type.Method'. preview=true returns a unified diff without writing.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in testInput) (*mcp.CallToolResult, any, error) {
-		if err := validateGoFile(in.File); err != nil {
-			return err, nil, nil
-		}
-		var testFile, diff string
-		var err error
-		if in.Preview {
-			diff, _, err = runPreview(ctx, commands, func(sc service.SurgeonCommands) error {
-				var innerErr error
-				testFile, innerErr = sc.GenerateTest(ctx, in.File, in.Identifier)
-				return innerErr
-			})
-		} else {
-			testFile, err = commands.GenerateTest(ctx, in.File, in.Identifier)
-		}
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("failed to generate test: %v", err), err), nil, nil
-		}
-		msg := fmt.Sprintf("SUCCESS: Generated test skeleton in %s", testFile)
-		if in.Preview {
-			msg = fmt.Sprintf("PREVIEW: would generate test skeleton in %s", testFile)
-			if diff != "" {
-				msg += "\n\n" + diff
-			}
-		}
-		res := textResult(msg)
-		res.StructuredContent = testOutput{TestFile: testFile, Identifier: in.Identifier}
-		return res, nil, nil
-	})
-
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "tag",
-		Description: "Manage struct field tags. Bulk: auto='json'/'bson' generates snake_case tags on all exported fields. Targeted: field + set updates one field's tag. patch_struct set_tag is an alternative for single fields. preview=true returns a unified diff without writing.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in tagInput) (*mcp.CallToolResult, any, error) {
-		if err := validateGoFile(in.File); err != nil {
-			return err, nil, nil
-		}
-		reqDomain := domain.TagRequest{
-			FilePath:   in.File,
-			StructName: in.Identifier,
-			FieldName:  in.Field,
-			SetTag:     in.Set,
-			AutoFormat: in.Auto,
-		}
-		var diff string
-		var err error
-		if in.Preview {
-			diff, _, err = runPreview(ctx, commands, func(sc service.SurgeonCommands) error {
-				return sc.TagStruct(ctx, reqDomain)
-			})
-		} else {
-			err = commands.TagStruct(ctx, reqDomain)
-		}
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("failed to update tags: %v", err), err), nil, nil
-		}
-		msg := fmt.Sprintf("SUCCESS: Updated tags for %s in %s", in.Identifier, in.File)
-		if in.Preview {
-			msg = fmt.Sprintf("PREVIEW (tag): %s in %s", in.Identifier, in.File)
-			if diff != "" {
-				msg += "\n\n" + diff
-			}
-		}
-		res := textResult(msg)
-		res.StructuredContent = tagOutput{File: in.File, Identifier: in.Identifier, Field: in.Field}
-		return res, nil, nil
-	})
-
-	mcp.AddTool(s, &mcp.Tool{
-		Name:        "extract_interface",
-		Description: "Derive an interface from a struct's exported methods. Use out to place it in a specific file. Set mock_file + mock_name to generate the mock in the same step. preview=true returns a unified diff without writing.",
-	}, func(ctx context.Context, req *mcp.CallToolRequest, in extractInterfaceInput) (*mcp.CallToolResult, any, error) {
-		if err := validateGoFile(in.File); err != nil {
-			return err, nil, nil
-		}
-		reqDomain := domain.ExtractInterfaceRequest{
-			FilePath:      in.File,
-			StructName:    in.Identifier,
-			InterfaceName: in.Name,
-			OutPath:       in.Out,
-			MockFile:      in.MockFile,
-			MockName:      in.MockName,
-		}
-		var interfaceFile, diff string
-		var err error
-		if in.Preview {
-			diff, _, err = runPreview(ctx, commands, func(sc service.SurgeonCommands) error {
-				var innerErr error
-				interfaceFile, innerErr = sc.ExtractInterface(ctx, reqDomain)
-				return innerErr
-			})
-		} else {
-			interfaceFile, err = commands.ExtractInterface(ctx, reqDomain)
-		}
-		if err != nil {
-			return errorResultWithCode(fmt.Sprintf("failed to extract interface: %v", err), err), nil, nil
-		}
-		msg := fmt.Sprintf("SUCCESS: Extracted interface %s into %s", in.Name, interfaceFile)
-		if in.Preview {
-			msg = fmt.Sprintf("PREVIEW: would extract interface %s into %s", in.Name, interfaceFile)
-			if diff != "" {
-				msg += "\n\n" + diff
-			}
-		}
-		res := textResult(msg)
-		res.StructuredContent = extractInterfaceOutput{InterfaceName: in.Name, InterfaceFile: interfaceFile, MockFile: in.MockFile, MockName: in.MockName}
-		return res, nil, nil
-	})
-}
-
-type executePlanActionInput struct {
-	Action            string                  `json:"action" jsonschema:"action type: create_file, replace_file, add_func, update_func, delete_func, add_struct, update_struct, delete_struct, add_interface, update_interface, delete_interface, insert_call, patch_function, patch_struct, patch_interface, patch_file, patch_decl"`
-	File              string                  `json:"file" jsonschema:"target file path"`
-	Package           string                  `json:"package,omitempty" jsonschema:"package import path (for cross-package operations)"`
-	Identifier        string                  `json:"identifier,omitempty" jsonschema:"AST identifier, e.g. FuncName or Receiver.Method"`
-	Content           string                  `json:"content,omitempty" jsonschema:"raw Go source code, no package declaration or imports"`
-	MockFile          string                  `json:"mock_file,omitempty" jsonschema:"target file for the generated mock"`
-	MockName          string                  `json:"mock_name,omitempty" jsonschema:"name of the mock struct"`
-	Doc               string                  `json:"doc,omitempty" jsonschema:"set or replace the doc comment (raw text without // prefix)"`
-	StripDoc          bool                    `json:"strip_doc,omitempty" jsonschema:"remove the existing doc comment"`
-	Position          string                  `json:"position,omitempty" jsonschema:"insert position: before-return, end-of-body, or after:<marker>"`
-	WithTest          bool                    `json:"with_test,omitempty" jsonschema:"generate a test skeleton alongside the function"`
-	PatchFunctionOps  []patchOpInput          `json:"patch_function_ops,omitempty" jsonschema:"patch operations for patch_function actions (same shape as patch_function tool's patches)"`
-	PatchStructOps    []structPatchOpInput    `json:"patch_struct_ops,omitempty" jsonschema:"patch operations for patch_struct actions"`
-	PatchInterfaceOps []interfacePatchOpInput `json:"patch_interface_ops,omitempty" jsonschema:"patch operations for patch_interface actions"`
-	PatchFileOps      []filePatchOpInput      `json:"patch_file_ops,omitempty" jsonschema:"patch operations for patch_file actions"`
-	PatchDeclOps      []patchOpInput          `json:"patch_decl_ops,omitempty" jsonschema:"patch operations for patch_decl actions (same shape as patch_function ops)"`
-}
-
-type testRunInput struct {
-	Dir              string   `json:"dir,omitempty" jsonschema:"directory to test (relative to the project root). Defaults to ./..."`
-	Run              string   `json:"run,omitempty" jsonschema:"optional -run regexp filter"`
-	Count            int      `json:"count,omitempty" jsonschema:"iterations per test (default 1)"`
-	Race             bool     `json:"race,omitempty" jsonschema:"enable the race detector"`
-	Tags             string   `json:"tags,omitempty" jsonschema:"build tags (whitelist [a-z_][a-z0-9_,.]*)"`
-	TimeoutSeconds   int      `json:"timeout_seconds,omitempty" jsonschema:"overall timeout in seconds (default 120, max 600)"`
-	AffectedBy       string   `json:"affected_by,omitempty" jsonschema:"path to a .go file — narrow the test run to the package that owns this file plus every package in the module that (transitively) imports it. Mutually exclusive with dir and symbols."`
-	Symbols          []string `json:"symbols,omitempty" jsonschema:"list of symbols in the form 'FuncName' or 'pkg.FuncName'. Each symbol is resolved to its owning package and a -run filter is built from Go naming conventions (^TestFuncName). Mutually exclusive with dir and affected_by. Use when you want to run only tests related to specific functions."`
-	Verbosity        string   `json:"verbosity,omitempty" jsonschema:"output verbosity: 'summary' returns only success, summary, and failed tests (compact ~1k chars regardless of suite size — ideal for large suites); 'full' returns everything including raw_output and per-test elapsed_ms. Defaults to auto: summary when the suite has more than 50 tests, full otherwise. Pass 'full' to force the verbose payload, 'summary' to force the compact one."`
-	IncludeRawOutput bool     `json:"include_raw_output,omitempty" jsonschema:"include the verbatim go test -json stream in structured output. By default RawOutput is dropped from the structured payload on success (where it bloats responses without adding signal); failures always keep it. Set true to force inclusion (e.g. when you want full test output regardless of pass/fail)."`
 }
